feat: refactor database imports and move constants to a new file
