+++ src/events/delete/onMessageDelete.ts
@@ -1,5 +1,5 @@
 import type { Client } from "discord.js";
-import db from "../../database.js";
+import db from "../../database";
 import { getTranslation } from "../../i18n";
 import { discordLogs } from "../../utils";
