+++ src/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from "./config";
+export * from "./constant";
 export * from "./menu";
